Added image vision to '/gpt'-command using ChatGPT-4 vision model (#260)
Added image vision capabilities to the '/gpt'-command. Now all images in the command message or any message in the same reply chain are sent with the request to openai as the context. As these images are fetched from Telegram server, works with images sent to Telegram before this update. System message can not contain an image.
- Switched openai library to plain http-requests for the gpt-command. Added max_tokens parameter with value of 4096 (maximum available for the vision model per single request)
- Updated gpt model catalogy. Added 'gpt-4-vision-preview', switched old 3.5 models to 'gpt-3.5-turbo-1106' which is newer upgrade.
- Bot uses requested major model version for the request. If any images, model is switched to best suited one with vision capabilities
- Added method to telethon_service for downloading all images associated with a message
- Upgraged and added tests. Updated mock_v2-classes to be able to handle messages with images